Suggest to update RPKI once an hour.
In autonomous systems running bgpd(8) and rpki-client(8) on their edge
routers, it may be beneficial when out-of-the-box all routers don't all
do rpki fetches & bgp loads at the same time. It is expected behavior
for RPKI information to un-evenly percolate towards the BGP edge in a
staggered way.

The 'once an hour' pace may be a reasonable balance between the needs of
internet users, and what network operators tolerate in churn.

OK deraadt@